Added the mechanical pump Updated version number (time to release, someday) Briefly added Emplacements Changed TMT a bit Effect Crates now work Added a new Debugger model Added Light Engineer Armor Added Chain Link Fence New models for ammo production machines New bullet models (currently unused) New sandbags models (currently unused)

+++ src/main/java/pl/pabilo8/immersiveintelligence/common/blocks/multiblocks/metal/tileentities/first/TileEntityRadioStation.java
@@ -310,7 +310,7 @@ public class TileEntityRadioStation extends TileEntityMultiblockMetal<TileEntity
 	@Override
 	public float getWeatherRangeDecrease()
 	{
-		return world.isRainingAt(getPos())?RadioStation.weatherHarshness: 1f;
+		return world.isRainingAt(getPos())?(float)RadioStation.weatherHarshness: 1f;
 	}
 
 	@Override
